backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment variables
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.0

# Date/Time utilities
python-dateutil==2.8.2

# Django filters
django-filter==23.5
